requirements.txt
Flask==2.3.3
pymongo==4.5.0
Werkzeug==2.3.7
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0
